requirements.txt
python-dotenv
pandas
numpy
ipykernel
matplotlib
seaborn
scikit-learn
librosa
pyaml
Pillow
## -e .